requirements.txt
absl-py==2.3.1
altair==5.5.0
attrs==25.3.0
blinker==1.9.0
brotlicffi @ file:///C:/b/abs_f4lixvzmxt/croot/brotlicffi_1736183286104/work
cachetools==6.2.0
certifi @ file:///C:/b/abs_2bz30lzohz/croot/certifi_1754570654565/work/certifi
cffi @ file:///C:/b/abs_29_b57if3f/croot/cffi_1736184144340/work
charset-normalizer @ file:///croot/charset-normalizer_1721748349566/work
click==8.1.8
colorama==0.4.6
contourpy==1.3.0
cycler==0.12.1
facenet-pytorch==2.6.0
filelock @ file:///C:/b/abs_758187j281/croot/filelock_1744281404850/work
flatbuffers==25.2.10
fonttools==4.59.2
fsspec==2025.7.0
gitdb==4.0.12
GitPython==3.1.45
gmpy2 @ file:///C:/b/abs_d8ki0o0h97/croot/gmpy2_1738085498525/work
grad-cam==1.5.5
huggingface-hub==0.34.4
idna @ file:///C:/b/abs_aad84bnnw5/croot/idna_1714398896795/work
importlib_metadata==8.7.0
importlib_resources==6.5.2
jax==0.4.30
jaxlib==0.4.30
Jinja2 @ file:///C:/b/abs_920kup4e6u/croot/jinja2_1741711580669/work
joblib==1.5.2
jsonschema==4.25.1
jsonschema-specifications==2025.4.1
kiwisolver==1.4.7
MarkupSafe @ file:///C:/b/abs_a0ma7ge0jc/croot/markupsafe_1738584052792/work
matplotlib==3.9.4
mediapipe==0.10.14
mkl-service==2.4.0
mkl_fft @ file:///C:/Users/dev-admin/mkl/mkl_fft_1730823082242/work
mkl_random @ file:///C:/Users/dev-admin/mkl/mkl_random_1730822522280/work
ml_dtypes==0.5.3
mpmath @ file:///C:/b/abs_7833jrbiox/croot/mpmath_1690848321154/work
narwhals==2.3.0
networkx @ file:///C:/b/abs_3bxnu56g9d/croot/networkx_1717597507456/work
numpy==1.26.4
opencv-contrib-python==4.12.0.88
opencv-python==4.12.0.88
opt_einsum==3.4.0
packaging==25.0
pandas==2.3.2
patsy==1.0.1
pillow==10.2.0
plotly==6.3.0
plotly-express==0.4.1
protobuf==4.25.8
pyarrow==21.0.0
pycparser @ file:///tmp/build/80754af9/pycparser_1636541352034/work
pydeck==0.9.1
pyparsing==3.2.3
PySocks @ file:///C:/ci/pysocks_1605307512533/work
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML @ file:///C:/b/abs_14xkfs39bx/croot/pyyaml_1728657968772/work
referencing==0.36.2
requests @ file:///C:/b/abs_f53smfvswb/croot/requests_1756709376206/work
rpds-py==0.27.1
safetensors==0.6.2
scikit-learn==1.6.1
scipy==1.13.1
seaborn==0.13.2
six==1.17.0
smmap==5.0.2
sounddevice==0.5.2
statsmodels==0.14.5
streamlit==1.49.1
streamlit-option-menu==0.4.0
sympy==1.13.1
tenacity==9.1.2
threadpoolctl==3.6.0
timm==1.0.19
toml==0.10.2
torch==2.2.2
torchaudio==2.5.1
torchvision==0.17.2
tornado==6.5.2
tqdm==4.67.1
ttach==0.0.3
typing_extensions @ file:///C:/b/abs_ecq8gc0vbm/croot/typing_extensions_1756281142218/work
tzdata==2025.2
urllib3 @ file:///C:/b/abs_393g1ayee_/croot/urllib3_1750775481226/work
watchdog==6.0.0
win-inet-pton @ file:///C:/ci/win_inet_pton_1605306162074/work
zipp==3.23.0
